New project/requirements.txt
fastapi>=0.115.0,<1.0.0
openai>=1.100.0,<2.0.0
pydantic>=2.9.0,<3.0.0
python-dotenv>=1.0.1,<2.0.0
requests>=2.32.0,<3.0.0
uvicorn>=0.30.0,<1.0.0
youtube-transcript-api>=1.2.3,<2.0.0
